# foundry/backend/openapi.yaml
openapi: 3.0.3
info:
  title: LabWired Foundry VaaS API
  description: |
    Verification-as-a-Service (VaaS) for autonomous AI hardware agents.
    Provides a standardized loop for task ingestion, context retrieval, and iterative formal verification.
  version: 1.1.0
servers:
  - url: http://localhost:8080/v1
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human/Agent-readable description of what went wrong.
        remediation:
          type: string
          description: Actionable suggestion for the calling agent.

    VerificationRequest:
      type: object
      required:
        - chip_yaml
      properties:
        chip_yaml:
          type: string
          description: LabWired-compatible YAML definition of the peripheral logic.

    SynthesisRequest:
      type: object
      required:
        - component_name
        - requirements
      properties:
        component_name:
          type: string
          description: Name of the component to synthesize (e.g., "BME280")
        requirements:
          type: string
          description: Natural language requirements or datasheet snippets describing the desired behavior.
        datasheet_url:
          type: string
          description: Optional URL to the component datasheet.

    VerificationResult:
      type: object
      properties:
        pass:
          type: boolean
        assertions_passed:
          type: integer
        assertions_total:
          type: integer
        compiler_logs:
          type: string
        vcd_url:
          type: string
          description: URL to download the Value Change Dump (VCD) trace.
        error:
          type: string
          description: Detailed error stack trace if verification failed.

  responses:
    UnauthorizedError:
      description: API key is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

paths:
  /info:
    get:
      summary: API Discovery
      description: Returns metadata about the Foundry engine capabilities.
      responses:
        '200':
          description: OK

  /estimate:
    post:
      summary: Estimate Synthesis Cost
      description: Get a quote for how many API runs a synthesis request will consume, based on complexity and novelty.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisRequest'
      responses:
        '200':
          description: Cost estimate.
          content:
            application/json:
              schema:
                type: object
                properties:
                  component_name:
                    type: string
                  estimated_cost_runs:
                    type: integer
                  message:
                    type: string

  /synthesize:
    post:
      summary: Autonomous Synthesis
      description: Submit high-level requirements. The Foundry's internal agents will autonomously write and formally verify a digital twin, returning the strict IR and proof.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisRequest'
      responses:
        '202':
          description: Synthesis job accepted. The process will run asynchronously and consume Quota.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string

  /models/verify:
    post:
      summary: Stateless Model Verification
      description: Submit a peripheral model for formal verification. Returns synchronous execution traces and compiler feedback.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        '200':
          description: Verification executed. Check `pass` boolean and `compiler_logs` for results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
  /systems/verify:
    post:
      summary: System-Level Verification
      description: Submit a multi-component system netlist for formal verification. Returns synchronous execution traces across multiple buses.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - system_yaml
              properties:
                system_yaml:
                  type: string
                  description: LabWired system YAML linking MCU and peripherals.
      responses:
        '200':
          description: System Verification executed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'